+++ adapter-meta/Sources/CloudXMetaAdapter/Initializers/CLXMetaInitializer.h
@@ -18,6 +18,7 @@
 + (NSString *)sdkVersion;
 
 - (void)initializeWithConfig:(nullable CLXBidderConfig *)config 
+                    testMode:(BOOL)testMode
                   completion:(void (^)(BOOL success, NSError * _Nullable error))completion;
 
 @end